requirements.txt
Django==4.2.5
gunicorn==20.1.0
djangorestframework==3.14.0